Remove reactions, move FileStore, move HtmlRoutes
