+++ sample/src/jvmMain/kotlin/external/Adapter.kt
@@ -16,7 +16,6 @@
 
 package external
 
-import io.gladed.watchable.store.FileStore
 import io.gladed.watchable.store.cached
 import logic.Logic
 import logic.Operations
