+++ sample/src/jvmTest/kotlin/AdapterTest.kt
@@ -17,8 +17,6 @@
 import external.Adapter
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.take
-import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.UnstableDefault
 import kotlinx.serialization.json.Json
@@ -52,24 +50,6 @@ class AdapterTest {
         runBlocking {
             Adapter.createLogic(coroutineContext, folder.root).scoped(this).apply {
                 assertEquals(robin, birds.get(robin.id))
-            }
-        }
-    }
-
-    @Test fun `modified chirp is stored`() {
-        runBlocking {
-            Adapter.createLogic(coroutineContext, folder.root).scoped(this).apply {
-                birds.put(robin.id, robin)
-                birds.put(wren.id, wren)
-                chirps.put(chirp.id, chirp)
-                chirp.reactions += wren.id to "+"
-            }
-        }
-
-        runBlocking {
-            Adapter.createLogic(coroutineContext, folder.root).scoped(this).apply {
-                val chirpId = chirps.keys().take(1).toList().first()
-                assertEquals(mapOf(wren.id to "+"), chirps.get(chirpId).reactions)
             }
         }
     }
